Improved Strideable Implementation & +() Bug Fix
I improved the Strideable implementation to cope with a lack of Int128
support being available, so in cases where integer arithmetic is not
needed the value will be accurate. Otherwise, revert to casting to
integer so that some action can be made. Also resolved a bug with the
precondition present in the +() operator.