Rough draft of Swift version 3.2/4.0 support
This is a major restructuring and in certain places rewriting of the
library. The Swift 3.2/4.0 UnsignedInteger changes are pretty massive
at the protocol layers. I might end up rebasing this later, as a lot of
source is ripped out in this version, but it was the only way I could
start making progress.

Skeletons for all required methods, protocols and tests should be
present. Most of the tests currently fail, as a number of the stub
methods needs to be fleshed out before most basic operation can work.