Bump version number to 0.8.0
As I'm nervous about how CocoaPods handles project tags, I want to
cut a new release specific to adding support for CocoaPods. This
release only has a single code change related to warnings in Xcode,
otherwise, it is just to play it safe when releasing the project
to CocoaPods' trunk.